Parametrize problem tests
The `pytest-lazy-fixture` library lets us use fixtures to parametrize
other fixtures, which makes it super easy to write a generic fixture
containing all valid problem fixtures we've written. This way, we can
add to that fixture as we write more problem fixtures, and we don't have
to write duplicate tests.