Improve student-facing ui (#38)
- Move cli stuff to directory
- Discover problems via files, not problem name
This seems much cleaner. For example, there's an awkward issue with the
other version where we want problems and submissions to have symbols
with the same name, but we couldn't do that when we required there to be
only one symbol with that name in the directory.

- Handle student-facing errors
This resolves #34.

- Move test runner facilities to a common module
This will make them much more reusable.